chore: use kres to setup build instructions
This modernizes the build system.
