+++ .conform.yaml
@@ -1,36 +1,39 @@
+# THIS FILE WAS AUTOMATICALLY GENERATED, PLEASE DO NOT EDIT.
+#
+# Generated on 2021-09-01T21:42:39Z by kres 5ca7963-dirty.
+
+---
 policies:
-  - type: commit
-    spec:
-      header:
-        length: 89
-        imperative: true
-        case: lower
-        invalidLastCharacters: .
-      body:
-        required: true
-      dco: true
-      gpg:
-        required: true
-        identity:
-          gitHubOrganization: talos-systems
-      spellcheck:
-        locale: US
-      maximumOfOneCommit: true
-      conventional:
-        types:
-          - chore
-          - docs
-          - perf
-          - refactor
-          - style
-          - test
-        scopes:
-          - policy
-  - type: license
-    spec:
-      includeSuffixes:
-        - .go
-      header: |
-        /* This Source Code Form is subject to the terms of the Mozilla Public
-         * License, v. 2.0. If a copy of the MPL was not distributed with this
-         * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+- type: commit
+  spec:
+    dco: true
+    gpg:
+      required: true
+      identity:
+        gitHubOrganization: talos-systems
+    spellcheck:
+      locale: US
+    maximumOfOneCommit: true
+    header:
+      length: 89
+      imperative: true
+      case: lower
+      invalidLastCharacters: .
+    body:
+      required: true
+    conventional:
+      types: ["chore","docs","perf","refactor","style","test","release"]
+      scopes: [".*"]
+- type: license
+  spec:
+    skipPaths:
+      - .git/
+    includeSuffixes:
+      - .go
+    excludeSuffixes:
+      - .pb.go
+      - .pb.gw.go
+    header: |
+      // This Source Code Form is subject to the terms of the Mozilla Public
+      // License, v. 2.0. If a copy of the MPL was not distributed with this
+      // file, You can obtain one at http://mozilla.org/MPL/2.0/.
