+++ .kres.yaml
@@ -1,0 +1,20 @@
+kind: golang.Build
+spec:
+  outputs:
+    conform-linux-amd64:
+      GOOS: linux
+      GOARCH: amd64
+    conform-linux-arm64:
+      GOOS: linux
+      GOARCH: arm64
+    conform-darwin-amd64:
+      GOOS: darwin
+      GOARCH: amd64
+    conform-darwin-arm64:
+      GOOS: darwin
+      GOARCH: arm64
+---
+kind: golang.Toolchain
+spec:
+  extraPackages:
+    - git
