+++ Dockerfile
@@ -1,40 +1,133 @@
-ARG GOLANG_IMAGE
-FROM ${GOLANG_IMAGE} AS common
+# syntax = docker/dockerfile-upstream:1.2.0-labs
+
+# THIS FILE WAS AUTOMATICALLY GENERATED, PLEASE DO NOT EDIT.
+#
+# Generated on 2021-09-01T21:55:41Z by kres 5ca7963-dirty.
+
+ARG TOOLCHAIN
+
+# cleaned up specs and compiled versions
+FROM scratch AS generate
+
+FROM ghcr.io/talos-systems/ca-certificates:v0.3.0-12-g90722c3 AS image-ca-certificates
+
+FROM ghcr.io/talos-systems/fhs:v0.3.0-12-g90722c3 AS image-fhs
+
+# runs markdownlint
+FROM node:14.8.0-alpine AS lint-markdown
+RUN npm i -g markdownlint-cli@0.23.2
+RUN npm i sentences-per-line@0.2.1
+WORKDIR /src
+COPY .markdownlint.json .
+COPY ./CHANGELOG.md ./CHANGELOG.md
+COPY ./README.md ./README.md
+RUN markdownlint --ignore "CHANGELOG.md" --ignore "**/node_modules/**" --ignore '**/hack/chglog/**' --rules /node_modules/sentences-per-line/index.js .
+
+# base toolchain image
+FROM ${TOOLCHAIN} AS toolchain
+RUN apk --update --no-cache add bash curl build-base protoc protobuf-dev git
+
+# build tools
+FROM toolchain AS tools
+ENV GO111MODULE on
 ENV CGO_ENABLED 0
-ENV GO111MODULES on
-WORKDIR /conform
-COPY go.mod ./
-COPY go.sum ./
-RUN go mod download
-RUN go mod verify
-COPY ./ ./
-RUN go list -mod=readonly all
+ENV GOPATH /go
+RUN curl -sfL https://install.goreleaser.com/github.com/golangci/golangci-lint.sh | bash -s -- -b /bin v1.41.1
+ARG GOFUMPT_VERSION
+RUN go install mvdan.cc/gofumpt/gofumports@${GOFUMPT_VERSION} \
+	&& mv /go/bin/gofumports /bin/gofumports
 
-FROM common AS build
-ARG TAG
+# tools and sources
+FROM tools AS base
+WORKDIR /src
+COPY ./go.mod .
+COPY ./go.sum .
+RUN --mount=type=cache,target=/go/pkg go mod download
+RUN --mount=type=cache,target=/go/pkg go mod verify
+COPY ./cmd ./cmd
+COPY ./internal ./internal
+RUN --mount=type=cache,target=/go/pkg go list -mod=readonly all >/dev/null
+
+# builds conform-darwin-amd64
+FROM base AS conform-darwin-amd64-build
+COPY --from=generate / /
+WORKDIR /src/cmd/conform
+ARG VERSION_PKG="github.com/talos-systems/conform/internal/version"
 ARG SHA
-ENV GOOS linux
-ENV GOARCH amd64
-RUN go build -o /conform-${GOOS}-${GOARCH} -ldflags "-s -w -X \"github.com/talos-systems/conform/cmd.Tag=${TAG}\" -X \"github.com/talos-systems/conform/cmd.SHA=${SHA}\"" .
-
 ARG TAG
+RUN --mount=type=cache,target=/root/.cache/go-build --mount=type=cache,target=/go/pkg GOARCH=amd64 GOOS=darwin go build -ldflags "-s -w -X ${VERSION_PKG}.Name=conform -X ${VERSION_PKG}.SHA=${SHA} -X ${VERSION_PKG}.Tag=${TAG}" -o /conform-darwin-amd64
+
+# builds conform-darwin-arm64
+FROM base AS conform-darwin-arm64-build
+COPY --from=generate / /
+WORKDIR /src/cmd/conform
+ARG VERSION_PKG="github.com/talos-systems/conform/internal/version"
 ARG SHA
-ENV GOOS darwin
-ENV GOARCH amd64
-RUN go build -o /conform-${GOOS}-${GOARCH} -ldflags "-s -w -X \"github.com/talos-systems/conform/cmd.Tag=${TAG}\" -X \"github.com/talos-systems/conform/cmd.SHA=${SHA}\"" .
+ARG TAG
+RUN --mount=type=cache,target=/root/.cache/go-build --mount=type=cache,target=/go/pkg GOARCH=arm64 GOOS=darwin go build -ldflags "-s -w -X ${VERSION_PKG}.Name=conform -X ${VERSION_PKG}.SHA=${SHA} -X ${VERSION_PKG}.Tag=${TAG}" -o /conform-darwin-arm64
 
-FROM common AS test
-ENV GOOS linux
-ENV GOARCH amd64
-RUN curl -sSfL https://raw.githubusercontent.com/golangci/golangci-lint/master/install.sh | sh -s -- -b $(go env GOPATH)/bin v1.42.0
-RUN golangci-lint run
-RUN CGO_ENABLED=1 go test -v -race -covermode=atomic -coverprofile=/coverage.txt ./...
+# builds conform-linux-amd64
+FROM base AS conform-linux-amd64-build
+COPY --from=generate / /
+WORKDIR /src/cmd/conform
+ARG VERSION_PKG="github.com/talos-systems/conform/internal/version"
+ARG SHA
+ARG TAG
+RUN --mount=type=cache,target=/root/.cache/go-build --mount=type=cache,target=/go/pkg GOARCH=amd64 GOOS=linux go build -ldflags "-s -w -X ${VERSION_PKG}.Name=conform -X ${VERSION_PKG}.SHA=${SHA} -X ${VERSION_PKG}.Tag=${TAG}" -o /conform-linux-amd64
 
-FROM alpine:3.11 as ca-certificates
-RUN apk add --update --no-cache ca-certificates
+# builds conform-linux-arm64
+FROM base AS conform-linux-arm64-build
+COPY --from=generate / /
+WORKDIR /src/cmd/conform
+ARG VERSION_PKG="github.com/talos-systems/conform/internal/version"
+ARG SHA
+ARG TAG
+RUN --mount=type=cache,target=/root/.cache/go-build --mount=type=cache,target=/go/pkg GOARCH=arm64 GOOS=linux go build -ldflags "-s -w -X ${VERSION_PKG}.Name=conform -X ${VERSION_PKG}.SHA=${SHA} -X ${VERSION_PKG}.Tag=${TAG}" -o /conform-linux-arm64
 
-FROM scratch AS image
-COPY --from=ca-certificates /etc/ssl/certs/ca-certificates.crt /etc/ssl/certs/
-COPY --from=build /conform-linux-amd64 /conform
-ENTRYPOINT [ "/conform" ]
-CMD [ "enforce" ]
+# runs gofumpt
+FROM base AS lint-gofumpt
+RUN find . -name '*.pb.go' | xargs -r rm
+RUN find . -name '*.pb.gw.go' | xargs -r rm
+RUN FILES="$(gofumports -l -local github.com/talos-systems/conform .)" && test -z "${FILES}" || (echo -e "Source code is not formatted with 'gofumports -w -local github.com/talos-systems/conform .':\n${FILES}"; exit 1)
+
+# runs golangci-lint
+FROM base AS lint-golangci-lint
+COPY .golangci.yml .
+ENV GOGC 50
+RUN --mount=type=cache,target=/root/.cache/go-build --mount=type=cache,target=/root/.cache/golangci-lint --mount=type=cache,target=/go/pkg golangci-lint run --config .golangci.yml
+
+# runs unit-tests with race detector
+FROM base AS unit-tests-race
+ARG TESTPKGS
+RUN --mount=type=cache,target=/root/.cache/go-build --mount=type=cache,target=/go/pkg --mount=type=cache,target=/tmp CGO_ENABLED=1 go test -v -race -count 1 ${TESTPKGS}
+
+# runs unit-tests
+FROM base AS unit-tests-run
+ARG TESTPKGS
+RUN --mount=type=cache,target=/root/.cache/go-build --mount=type=cache,target=/go/pkg --mount=type=cache,target=/tmp go test -v -covermode=atomic -coverprofile=coverage.txt -coverpkg=${TESTPKGS} -count 1 ${TESTPKGS}
+
+FROM scratch AS conform-darwin-amd64
+COPY --from=conform-darwin-amd64-build /conform-darwin-amd64 /conform-darwin-amd64
+
+FROM scratch AS conform-darwin-arm64
+COPY --from=conform-darwin-arm64-build /conform-darwin-arm64 /conform-darwin-arm64
+
+FROM scratch AS conform-linux-amd64
+COPY --from=conform-linux-amd64-build /conform-linux-amd64 /conform-linux-amd64
+
+FROM scratch AS conform-linux-arm64
+COPY --from=conform-linux-arm64-build /conform-linux-arm64 /conform-linux-arm64
+
+FROM scratch AS unit-tests
+COPY --from=unit-tests-run /src/coverage.txt /coverage.txt
+
+FROM conform-linux-${TARGETARCH} AS conform
+
+FROM scratch AS image-conform
+ARG TARGETARCH
+COPY --from=conform conform-linux-${TARGETARCH} /conform
+COPY --from=image-fhs / /
+COPY --from=image-ca-certificates / /
+LABEL org.opencontainers.image.source https://github.com/talos-systems/conform
+ENTRYPOINT ["/conform"]
+
