+++ cmd/conform/serve.go
@@ -1,8 +1,8 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-package cmd
+package main
 
 import (
 	"encoding/json"
@@ -50,7 +50,7 @@ var serveCmd = &cobra.Command{
 
 					return
 				}
-				// nolint: errcheck
+				//nolint: errcheck
 				defer os.RemoveAll(dir)
 
 				if err = os.MkdirAll(filepath.Join(dir, "github"), 0o700); err != nil {
