+++ cmd/conform/version.go
@@ -1,0 +1,32 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+package main
+
+import (
+	"github.com/spf13/cobra"
+
+	"github.com/talos-systems/conform/internal/version"
+)
+
+var shortVersion bool
+
+// versionCmd represents the version command.
+var versionCmd = &cobra.Command{
+	Use:   "version",
+	Short: "Prints the version",
+	Long:  ``,
+	Run: func(cmd *cobra.Command, args []string) {
+		if shortVersion {
+			version.PrintShortVersion()
+		} else {
+			version.PrintLongVersion()
+		}
+	},
+}
+
+func init() {
+	versionCmd.Flags().BoolVar(&shortVersion, "short", false, "Print the short version")
+	rootCmd.AddCommand(versionCmd)
+}
