+++ internal/policy/commit/check_conventional_commit.go
@@ -1,6 +1,6 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 package commit
 
@@ -61,7 +61,6 @@ func (c ConventionalCommitCheck) Errors() []error {
 }
 
 // ValidateConventionalCommit returns the commit type.
-//nolint:gocyclo,cyclop
 func (c Commit) ValidateConventionalCommit() policy.Check {
 	check := &ConventionalCommitCheck{}
 	groups := parseHeader(c.msg)
