+++ internal/policy/commit/check_gpg_identity.go
@@ -1,6 +1,6 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 package commit
 
@@ -19,6 +19,8 @@ import (
 )
 
 // GPGIdentityCheck ensures that the commit is cryptographically signed using known identity.
+//
+//nolint:govet
 type GPGIdentityCheck struct {
 	errors   []error
 	identity string
