+++ internal/policy/commit/check_header_length.go
@@ -1,6 +1,6 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 package commit
 
@@ -18,6 +18,8 @@ var MaxNumberOfCommitCharacters = 89
 
 // HeaderLengthCheck enforces a maximum number of charcters on the commit
 // header.
+//
+//nolint:govet
 type HeaderLengthCheck struct {
 	headerLength int
 	errors       []error
