+++ internal/policy/commit/check_number_of_commits.go
@@ -1,6 +1,6 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 package commit
 
@@ -15,6 +15,8 @@ import (
 
 // NumberOfCommits enforces a maximum number of charcters on the commit
 // header.
+//
+//nolint:govet
 type NumberOfCommits struct {
 	ref    string
 	ahead  int
